shorten classpath argument for storm processes
Instead of a fully-qualified and expanded classpath, we leverage Java 6's "wildcard
classpath" support.

Notably, just changing bin/storm to use a wildcard for its -cp argument was insufficient
for getting the worker classpath arguments to be short.  This was because the supervisor
looks at its own `java.class.path` property, which the JVM has auto-expanded from the
wildcard path into the set of JARs found at the wildcard location.  So we add a mechanism
to pass the original classpath into the supervisor, to be used as the basis for the
worker's classpath.